+++ hmf-common/src/main/java/com/hartwig/hmftools/common/genome/chromosome/ChromosomeLength.java
@@ -11,5 +11,5 @@ public interface ChromosomeLength {
     @NotNull
     String chromosome();
 
-    long length();
+    int length();
 }
