+++ hmf-common/src/main/java/com/hartwig/hmftools/common/genome/position/GenomePosition.java
@@ -9,7 +9,7 @@ public interface GenomePosition extends Comparable<GenomePosition> {
     @NotNull
     String chromosome();
 
-    long position();
+    int position();
 
     @Override
     default int compareTo(@NotNull GenomePosition other) {
